Split GenDialects up into smaller files
The main goal here is to make the distinction between the "model"
(description of dialects, operations, etc.) and the "view/printer"
(output of generated C++ code) clearer.
