+++ lib/TableGen/GenDialect.cpp
@@ -14,9 +14,14 @@
  * limitations under the License. 
  */
 
-#include "llvm-dialects/TableGen/Common.h"
-#include "llvm-dialects/TableGen/Format.h"
 #include "llvm-dialects/TableGen/GenDialect.h"
+#include "llvm-dialects/TableGen/Common.h"
+#include "llvm-dialects/TableGen/Constraints.h"
+#include "llvm-dialects/TableGen/Dialects.h"
+#include "llvm-dialects/TableGen/Format.h"
+#include "llvm-dialects/TableGen/Operations.h"
+#include "llvm-dialects/TableGen/Predicates.h"
+#include "llvm-dialects/TableGen/Traits.h"
 
 #include "llvm/ADT/StringExtras.h"
 #include "llvm/Support/CommandLine.h"
@@ -31,539 +36,6 @@ using namespace llvm;
 namespace {
 
 cl::opt<std::string> g_dialect("dialect", cl::desc("the dialect to generate"), cl::init(""));
-
-class BuiltinType;
-class Constraint;
-class Dialect;
-class OpClass;
-struct OpNamedValue;
-class Operation;
-class PredicateExpr;
-class Trait;
-
-class DialectsContext {
-public:
-  void init(RecordKeeper& records, const DenseSet<StringRef>& dialects);
-
-  Trait* getTrait(Record* traitRec);
-  Constraint* getConstraint(Record* constraintRec);
-  OpClass* getOpClass(Record* opClassRec);
-  Dialect* getDialect(Record* dialectRec);
-
-  BuiltinType* getVoidTy() const {return m_voidTy;}
-
-  std::unique_ptr<PredicateExpr> parsePredicateExpr(DagInit *dag);
-
-private:
-  std::vector<OpNamedValue> parseArguments(Record* rec);
-
-  BuiltinType* m_voidTy = nullptr;
-  DenseMap<Record*, std::unique_ptr<Trait>> m_traits;
-  DenseMap<Record*, std::unique_ptr<Constraint>> m_constraints;
-  DenseMap<Record*, std::unique_ptr<OpClass>> m_opClasses;
-  DenseMap<Record*, std::unique_ptr<Dialect>> m_dialects;
-};
-
-class Constraint {
-public:
-  enum class Kind : uint8_t {
-    Type_First,
-    BuiltinType = Type_First,
-    DialectType,
-    Type_Last = DialectType,
-
-    Attr,
-    BaseCPred_First,
-    BaseCPred = BaseCPred_First,
-    SameTypes,
-    BaseCPred_Last = SameTypes,
-  };
-
-  virtual ~Constraint() = default;
-
-  virtual void init(DialectsContext* context, Record* record) {
-    m_record = record;
-    m_builderArgumentFilter = record->getValueAsString("builderArgumentFilter");
-  }
-
-  Kind getKind() const {return m_kind;}
-  Record* getRecord() const {return m_record;}
-  StringRef getName() const {return m_record->getName();}
-  StringRef getCppType() const;
-
-  StringRef getBuilderArgumentFilter() const {return m_builderArgumentFilter;}
-
-  virtual std::pair<unsigned, unsigned> getMinMaxArgs() const = 0;
-
-  /// Return a string containing a C++ expression that applies this predicate
-  /// to the given arguments.
-  virtual std::string
-  apply(FmtContext *fmt, ArrayRef<StringRef> arguments) const = 0;
-
-protected:
-  Constraint(Kind kind) : m_kind(kind) {}
-
-private:
-  const Kind m_kind;
-  Record* m_record = nullptr;
-  std::string m_builderArgumentFilter;
-};
-
-class Type : public Constraint {
-public:
-  static bool classof(const Constraint* c) {
-    return c->getKind() >= Kind::Type_First && c->getKind() <= Kind::Type_Last;
-  }
-
-  void init(DialectsContext* context, Record* record) override {
-    Constraint::init(context, record);
-  }
-
-  std::pair<unsigned, unsigned> getMinMaxArgs() const final {return {1, 1};}
-
-  std::string
-  apply(FmtContext *fmt, ArrayRef<StringRef> arguments) const final {
-    assert(arguments.size() == 1);
-    return tgfmt("$0 == $1", fmt, arguments[0], getLlvmType(fmt));
-  }
-
-  /// Return a string containing a C++ expression that returns the `llvm::Type*`
-  /// for this type.
-  virtual std::string getLlvmType(FmtContext *fmt) const = 0;
-
-protected:
-  Type(Kind kind) : Constraint(kind) {}
-};
-
-class BuiltinType : public Type {
-public:
-  BuiltinType() : Type(Kind::BuiltinType) {}
-
-  void init(DialectsContext* context, Record* record) override {
-    Type::init(context, record);
-    m_getter = record->getValueAsString("getter");
-  }
-
-  StringRef getGetter() const {return m_getter;}
-
-  static bool classof(const Constraint* c) {return c->getKind() == Kind::BuiltinType;}
-
-  std::string getLlvmType(FmtContext *fmt) const final {
-    return tgfmt(getGetter(), fmt);
-  }
-
-private:
-  std::string m_getter;
-};
-
-class DialectType : public Type {
-public:
-  DialectType() : Type(Kind::DialectType) {}
-
-  void init(DialectsContext* context, Record* record) override {
-    Type::init(context, record);
-
-    m_dialectRec = record->getValueAsDef("dialect");
-    if (!m_dialectRec->isSubClassOf("Dialect")) {
-      report_fatal_error(Twine("'dialect' field of type constraint '")
-                              + record->getName() + "' is not a subclass of Dialect");
-    }
-    m_mnemonic = record->getValueAsString("mnemonic");
-  }
-
-  Record* getDialectRec() const {return m_dialectRec;}
-  StringRef getMnemonic() const {return m_mnemonic;}
-
-  static bool classof(const Constraint* c) {return c->getKind() == Kind::DialectType;}
-
-  std::string getLlvmType(FmtContext *fmt) const final {
-    return tgfmt("$0::get($_builder)", fmt, getName());
-  }
-
-private:
-  Record* m_dialectRec = nullptr;
-  std::string m_mnemonic;
-};
-
-class Attr : public Constraint {
-public:
-  Attr() : Constraint(Kind::Attr) {}
-
-  void init(DialectsContext* context, Record* record) override {
-    Constraint::init(context, record);
-
-    m_cppType = record->getValueAsString("cppType");
-    Record* llvmTypeRec = record->getValueAsDef("llvmType");
-    Constraint* llvmType = context->getConstraint(llvmTypeRec);
-    if (!isa<Type>(llvmType)) {
-      report_fatal_error(Twine("Attr '") + record->getName()
-                              + "' has llvmType '" + llvmType->getName()
-                              + "' which is not a Type");
-    }
-
-    m_llvmType = cast<Type>(llvmType);
-    m_toLlvmValue = record->getValueAsString("toLlvmValue");
-    m_fromLlvmValue = record->getValueAsString("fromLlvmValue");
-  }
-
-  std::pair<unsigned, unsigned> getMinMaxArgs() const final {return {1, 1};}
-
-  std::string
-  apply(FmtContext *fmt, ArrayRef<StringRef> arguments) const final {
-    llvm_unreachable("cannot apply an Attr predicate");
-  }
-
-  StringRef getCppType() const {return m_cppType;}
-  Type* getLlvmType() const {return m_llvmType;}
-  StringRef getToLlvmValue() const {return m_toLlvmValue;}
-  StringRef getFromLlvmValue() const {return m_fromLlvmValue;}
-
-  static bool classof(const Constraint* c) {return c->getKind() == Kind::Attr;}
-
-private:
-  std::string m_cppType;
-  Type* m_llvmType = nullptr;
-  std::string m_toLlvmValue;
-  std::string m_fromLlvmValue;
-};
-
-class BaseCPred : public Constraint {
-public:
-  BaseCPred(StringRef name) : Constraint(getBaseCPredKind(name)) {}
-
-  static bool classof(const Constraint* c) {
-    return c->getKind() >= Kind::BaseCPred_First &&
-           c->getKind() <= Kind::BaseCPred_Last;
-  }
-
-  void init(DialectsContext* context, Record* record) override {
-    Constraint::init(context, record);
-
-    m_predExpr = record->getValueAsString("predExpr");
-
-    DagInit *arguments = record->getValueAsDag("arguments");
-    if (arguments->getOperatorAsDef({})->getName() != "ins") {
-      report_fatal_error(Twine("BaseCPred '") + record->getName()
-                             + "': arguments dag operator must be 'ins'");
-    }
-
-    for (auto [argInit, argName]
-             : llvm::zip(arguments->getArgs(), arguments->getArgNames())) {
-      if (m_variadic) {
-        report_fatal_error(Twine("BaseCPred '") + record->getName()
-                               + "': seq must be last");
-      }
-
-      Argument arg;
-      arg.name = argName->getValue();
-
-      if (arg.name.empty()) {
-        report_fatal_error(Twine("BaseCPred '") + record->getName()
-                               + "': missing argument name");
-      }
-
-      if (isa<UnsetInit>(argInit)) {
-        m_arguments.push_back(arg);
-      } else {
-        if (!isa<DefInit>(argInit) ||
-            cast<DefInit>(argInit)->getAsString() != "seq") {
-          report_fatal_error(Twine("BaseCPred '") + record->getName()
-                                 + "': bad def used in arguments");
-        }
-
-        m_variadic = arg;
-      }
-    }
-  }
-
-  std::pair<unsigned, unsigned> getMinMaxArgs() const final {
-    unsigned min = m_arguments.size();
-    return {min, m_variadic ? std::numeric_limits<unsigned>::max() : min};
-  }
-
-  std::string
-  apply(FmtContext *fmt, ArrayRef<StringRef> arguments) const final {
-    assert(arguments.size() >= m_arguments.size());
-    assert(m_variadic || arguments.size() == m_arguments.size());
-
-    FmtContextScope scope(*fmt);
-    for (const auto &[formal, actual] : llvm::zip(m_arguments, arguments)) {
-      if (formal.name == "_self")
-        fmt->withSelf(actual);
-      else
-        fmt->addSubst(formal.name, actual);
-    }
-
-    if (m_variadic) {
-      std::string actuals;
-      actuals += '{';
-
-      for (const auto &actual : arguments.drop_front(m_arguments.size())) {
-        if (actuals.size() > 1)
-          actuals += ", ";
-        actuals += actual;
-      }
-
-      actuals += '}';
-      fmt->addSubst(m_variadic->name, actuals);
-    }
-
-    return tgfmt(m_predExpr, fmt);
-  }
-
-private:
-  static Kind getBaseCPredKind(StringRef name) {
-    if (name == "SameTypes")
-      return Kind::SameTypes;
-    return Kind::BaseCPred;
-  }
-
-  struct Argument {
-    std::string name;
-  };
-
-  SmallVector<Argument> m_arguments;
-  std::optional<Argument> m_variadic;
-  std::string m_predExpr;
-};
-
-class PredicateExpr {
-public:
-  enum class Kind {
-    Logic_First,
-    And = Logic_First,
-    Or,
-    Not,
-    Logic_Last = Not,
-    Apply,
-  };
-
-  virtual ~PredicateExpr() = default;
-
-  Kind getKind() const {return m_kind;}
-
-  /// Return a string containing a C++ expression that evaluates this
-  /// predicate expression.
-  ///
-  /// @p names maps TableGen argument names (such as `$lhs`) to C++ names.
-  virtual std::string
-  evaluate(FmtContext *fmt,
-           const DenseMap<StringRef, std::string> &names) const = 0;
-
-protected:
-  PredicateExpr(Kind kind) : m_kind(kind) {}
-
-  const Kind m_kind;
-};
-
-class PredicateLogic : public PredicateExpr {
-public:
-  PredicateLogic(Kind kind,
-                 MutableArrayRef<std::unique_ptr<PredicateExpr>> arguments)
-      : PredicateExpr(kind) {
-    assert(classof(this));
-
-    for (auto &arg : arguments)
-      m_arguments.push_back(std::move(arg));
-  }
-
-  static bool classof(const PredicateExpr *e) {
-    return e->getKind() >= Kind::Logic_First &&
-           e->getKind() <= Kind::Logic_Last;
-  }
-
-  std::string
-  evaluate(FmtContext *fmt,
-           const DenseMap<StringRef, std::string> &names) const final {
-    if (getKind() == Kind::Not)
-      return "!" + m_arguments[0]->evaluate(fmt, names);
-
-    std::string out;
-    out += '(';
-
-    for (const auto &argument : m_arguments) {
-      if (out.size() > 1) {
-        if (getKind() == Kind::And)
-          out += " && ";
-        else
-          out += " || ";
-      }
-
-      out += argument->evaluate(fmt, names);
-    }
-
-    out += ')';
-    return out;
-  }
-
-private:
-  std::vector<std::unique_ptr<PredicateExpr>> m_arguments;
-};
-
-class PredicateApply : public PredicateExpr {
-public:
-  PredicateApply(Constraint *constraint, ArrayRef<std::string> arguments)
-      : PredicateExpr(Kind::Apply), m_constraint(constraint)
-      , m_arguments(arguments) {
-    assert(m_arguments.size() >= m_constraint->getMinMaxArgs().first);
-    assert(m_arguments.size() <= m_constraint->getMinMaxArgs().second);
-  }
-
-  static bool classof(const PredicateExpr *e) {
-    return e->getKind() == Kind::Apply;
-  }
-
-  Constraint *getPredicate() const {return m_constraint;}
-  ArrayRef<std::string> arguments() const {return m_arguments;}
-
-  std::string
-  evaluate(FmtContext *fmt,
-           const DenseMap<StringRef, std::string> &names) const final {
-    SmallVector<StringRef> arguments;
-    for (StringRef arg : m_arguments) {
-      auto it = names.find(arg);
-      if (it == names.end()) {
-        report_fatal_error(Twine("Evaluating ") + getAsTableGenString()
-                               + ": argument '" + arg + "' not found");
-      }
-      arguments.push_back(it->second);
-    }
-    return '(' + m_constraint->apply(fmt, arguments) + ')';
-  }
-
-  std::string getAsTableGenString() const {
-    std::string result;
-    result += '(';
-    result += m_constraint->getRecord()->getName();
-
-    for (const auto &arg : m_arguments) {
-      result += ' ';
-      result += arg;
-    }
-
-    result += ')';
-    return result;
-  }
-
-private:
-  Constraint *m_constraint;
-  std::vector<std::string> m_arguments;
-};
-
-class Trait {
-public:
-  enum class Kind : uint8_t {
-    LlvmAttributeTrait,
-  };
-
-  virtual ~Trait() {}
-
-  virtual void init(DialectsContext* context, Record* record) {
-    m_record = record;
-  }
-
-  Kind getKind() const {return m_kind;}
-  Record* getRecord() const {return m_record;}
-  StringRef getName() const {return m_record->getName();}
-
-protected:
-  Trait(Kind kind) : m_kind(kind) {}
-
-private:
-  const Kind m_kind;
-  Record* m_record = nullptr;
-};
-
-class LlvmAttributeTrait : public Trait {
-public:
-  LlvmAttributeTrait() : Trait(Kind::LlvmAttributeTrait) {}
-
-  void init(DialectsContext* context, Record* record) override {
-    Trait::init(context, record);
-    m_llvmEnum = record->getValueAsString("llvmEnum");
-  }
-
-  StringRef getLlvmEnum() const {return m_llvmEnum;}
-
-  static bool classof(const Trait* t) {return t->getKind() == Kind::LlvmAttributeTrait;}
-
-private:
-  std::string m_llvmEnum;
-};
-
-struct OpNamedValue {
-  std::string name;
-  Constraint* type = nullptr;
-};
-
-struct OverloadKey {
-  enum Kind {
-    Result,
-    Argument,
-  };
-
-  Kind kind;
-  unsigned index;
-};
-
-class OpClass {
-public:
-  OpClass* superclass = nullptr;
-  std::string name;
-  std::vector<OpNamedValue> arguments; // does not explicitly list superclass arguments
-  std::vector<OpClass*> subclasses;
-  std::vector<Operation*> operations;
-
-  SmallVector<OpNamedValue> getFullArguments() const {
-    SmallVector<OpNamedValue> args;
-    if (superclass)
-      args = superclass->getFullArguments();
-    args.insert(args.end(), arguments.begin(), arguments.end());
-    return args;
-  }
-  unsigned getNumFullArguments() const {
-    if (superclass)
-      return superclass->getNumFullArguments() + arguments.size();
-    return arguments.size();
-  }
-};
-
-class Operation {
-public:
-  OpClass* superclass = nullptr;
-  std::string name;
-  std::string mnemonic;
-  std::vector<Trait*> traits;
-  std::vector<OpNamedValue> arguments; // does not explicitly list superclass arguments
-  std::vector<OpNamedValue> results;
-  std::vector<std::unique_ptr<PredicateExpr>> verifier;
-  std::vector<OverloadKey> overloadKeys;
-  bool builderHasExplicitResultTypes = false;
-
-  SmallVector<OpNamedValue> getFullArguments() const {
-    SmallVector<OpNamedValue> args;
-    if (superclass)
-      args = superclass->getFullArguments();
-    args.insert(args.end(), arguments.begin(), arguments.end());
-    return args;
-  }
-  unsigned getNumFullArguments() const {
-    if (superclass)
-      return superclass->getNumFullArguments() + arguments.size();
-    return arguments.size();
-  }
-};
-
-class Dialect {
-public:
-  Record* record;
-  std::string cppName;
-  std::string name;
-  std::string cppNamespace;
-  std::vector<DialectType*> types;
-  std::vector<OpClass*> opClasses;
-  std::vector<std::unique_ptr<Operation>> operations;
-};
 
 /// Helper class for choosing unique variable names.
 class SymbolTable {
@@ -624,317 +96,12 @@ StringRef Constraint::getCppType() const {
   return "::llvm::Value *";
 }
 
-Trait* DialectsContext::getTrait(Record* traitRec) {
-  if (!traitRec->isSubClassOf("Trait"))
-    report_fatal_error(Twine("Trying to use '") + traitRec->getName()
-                           + "' as a constraint, but it is not a subclass of 'Constraint'");
-
-  auto& result = m_traits[traitRec];
-  if (!result) {
-    if (traitRec->isSubClassOf("LlvmAttributeTrait")) {
-      result = std::make_unique<LlvmAttributeTrait>();
-    } else {
-      report_fatal_error(Twine("unsupported trait: ") + traitRec->getName());
-    }
-    result->init(this, traitRec);
-  }
-  return result.get();
-}
-
-Constraint* DialectsContext::getConstraint(Record* constraintRec) {
-  if (!constraintRec->isSubClassOf("Constraint"))
-    report_fatal_error(Twine("Trying to use '") + constraintRec->getName()
-                           + "' as a constraint, but it is not a subclass of 'Constraint'");
-
-  auto& result = m_constraints[constraintRec];
-  if (!result) {
-    if (constraintRec->isSubClassOf("BuiltinType")) {
-      result = std::make_unique<BuiltinType>();
-    } else if (constraintRec->isSubClassOf("DialectType")) {
-      result = std::make_unique<DialectType>();
-    } else if (constraintRec->isSubClassOf("Attr")) {
-      result = std::make_unique<Attr>();
-    } else if (constraintRec->isSubClassOf("BaseCPred")) {
-      result = std::make_unique<BaseCPred>(constraintRec->getName());
-    } else {
-      report_fatal_error(Twine("unsupported type constraint: ") + constraintRec->getName());
-    }
-
-    result->init(this, constraintRec);
-  }
-  return result.get();
-}
-
-Dialect* DialectsContext::getDialect(Record* dialectRec) {
-  if (!dialectRec->isSubClassOf("Dialect"))
-    report_fatal_error(Twine("Trying to use '") + dialectRec->getName()
-                           + "' as a dialect, but it is not a subclass of 'Dialect'");
-
-  auto it = m_dialects.find(dialectRec);
-  if (it == m_dialects.end())
-    report_fatal_error(Twine("Trying to use dialect '") + dialectRec->getName()
-                           + "', but has not been initialized");
-
-  return it->second.get();
-}
-
-std::vector<OpNamedValue> DialectsContext::parseArguments(Record* rec) {
-  Record* superClassRec = rec->getValueAsDef("superclass");
-  OpClass* superclass = getOpClass(superClassRec);
-  DagInit* argsInit = rec->getValueAsDag("arguments");
-  std::vector<OpNamedValue> arguments;
-
-  if (argsInit->getOperatorAsDef({})->getName() != "ins")
-    report_fatal_error(Twine(rec->getName()) + " argument operator must be 'ins'");
-
-  for (unsigned i = 0; i < argsInit->getNumArgs(); ++i) {
-    if (superclass && i == 0) {
-      if (argsInit->getArgName(0) ||
-          argsInit->getArg(0) != superClassRec->getDefInit())
-        report_fatal_error(Twine(rec->getName()) + ": superclass must be first in arguments list");
-      continue;
-    }
-
-    OpNamedValue opArg;
-    opArg.name = argsInit->getArgNameStr(i);
-    if (auto* arg = dyn_cast_or_null<DefInit>(argsInit->getArg(i)))
-      opArg.type = getConstraint(arg->getDef());
-    if (!opArg.type) {
-      report_fatal_error(Twine(rec->getName()) + " argument " + Twine(i) +
-                            ": bad type constraint");
-    }
-    arguments.push_back(std::move(opArg));
-  }
-
-  return arguments;
-}
-
-OpClass* DialectsContext::getOpClass(Record* opClassRec) {
-  if (opClassRec->getName() == "NoSuperClass")
-    return nullptr;
-
-  if (!opClassRec->isSubClassOf("OpClass"))
-    report_fatal_error(Twine("Trying to use '") + opClassRec->getName()
-                           + "' as operation class, but it is not a subclass of 'OpClass'");
-
-  auto it = m_opClasses.find(opClassRec);
-  if (it != m_opClasses.end()) {
-    if (!it->second)
-      report_fatal_error(Twine("Circular superclass relations involving '")
-                             + opClassRec->getName() + "'");
-    return it->second.get();
-  }
-
-  // Leave a marker in the map to detect recursive superclass relations.
-  m_opClasses.try_emplace(opClassRec);
-
-  Dialect* dialect = getDialect(opClassRec->getValueAsDef("dialect"));
-  auto opClass = std::make_unique<OpClass>();
-  opClass->name = opClassRec->getName();
-  opClass->superclass = getOpClass(opClassRec->getValueAsDef("superclass"));
-  opClass->arguments = parseArguments(opClassRec);
-
-  OpClass* ptr = opClass.get();
-  if (opClass->superclass)
-    opClass->superclass->subclasses.push_back(ptr);
-  dialect->opClasses.push_back(ptr);
-  m_opClasses.find(opClassRec)->second = std::move(opClass);
-  return ptr;
-}
-
-void DialectsContext::init(RecordKeeper& records,
-                           const DenseSet<StringRef>& dialects) {
-  m_voidTy = cast<BuiltinType>(getConstraint(records.getDef("VoidTy")));
-
-  for (Record* dialectRec : records.getAllDerivedDefinitions("Dialect")) {
-    auto name = dialectRec->getValueAsString("name");
-    if (!dialects.contains(name))
-      continue;
-
-    auto dialect = std::make_unique<Dialect>();
-    dialect->record = dialectRec;
-    dialect->cppName = dialectRec->getName();
-    dialect->name = name;
-    dialect->cppNamespace = dialectRec->getValueAsString("cppNamespace");
-    m_dialects.insert(std::make_pair(dialectRec, std::move(dialect)));
-  }
-
-  for (Record* typeRec : records.getAllDerivedDefinitions("DialectType")) {
-    auto* dialectType = cast<DialectType>(getConstraint(typeRec));
-    auto dialectIt = m_dialects.find(dialectType->getDialectRec());
-    if (dialectIt != m_dialects.end())
-      dialectIt->second->types.push_back(dialectType);
-  }
-
-  for (Record* opRec : records.getAllDerivedDefinitions("Op")) {
-    Record* dialectRec = opRec->getValueAsDef("dialect");
-    auto dialectIt = m_dialects.find(dialectRec);
-    if (dialectIt == m_dialects.end())
-      continue;
-
-    // Extract information from the TableGen record.
-    auto op = std::make_unique<Operation>();
-    op->superclass = getOpClass(opRec->getValueAsDef("superclass"));
-    if (op->superclass)
-      op->superclass->operations.push_back(op.get());
-    op->name = opRec->getName();
-    op->mnemonic = opRec->getValueAsString("mnemonic");
-    for (Record* traitRec : opRec->getValueAsListOfDefs("traits"))
-      op->traits.push_back(getTrait(traitRec));
-
-    op->arguments = parseArguments(opRec);
-
-    DagInit* results = opRec->getValueAsDag("results");
-    assert(results->getOperatorAsDef({})->getName() == "outs");
-    assert(results->getNumArgs() <= 1 && "multiple result values not supported");
-    for (unsigned i = 0; i < results->getNumArgs(); ++i) {
-      OpNamedValue opResult;
-      opResult.name = results->getArgNameStr(i);
-      if (auto* result = dyn_cast_or_null<DefInit>(results->getArg(i)))
-        opResult.type = getConstraint(result->getDef());
-      if (!opResult.type) {
-        report_fatal_error(Twine("Operation '") + op->mnemonic + "' result " + Twine(i) +
-                               ": bad type constraint");
-      }
-      if (!isa<Attr>(opResult.type) && !isa<Type>(opResult.type))
-        op->builderHasExplicitResultTypes = true;
-      op->results.push_back(std::move(opResult));
-    }
-
-    ListInit *verifier = opRec->getValueAsListInit("verifier");
-    for (Init *ruleInit : *verifier) {
-      auto *rule = dyn_cast<DagInit>(ruleInit);
-      if (!rule) {
-        report_fatal_error(Twine("Operation '") + op->mnemonic
-                               + "': verifier rules must be dags");
-      }
-
-      op->verifier.push_back(parsePredicateExpr(rule));
-    }
-
-    // Derive the overload keys: Scan through results and arguments. Whenever
-    // we encounter one whose type isn't fully specified, add it to the overload
-    // keys unless an equal type has already been added.
-    auto needsOverloadKey = [&](const OpNamedValue &namedValue) -> bool {
-      if (isa<Type>(namedValue.type) || isa<Attr>(namedValue.type))
-        return false;
-
-      for (const auto &expr : op->verifier) {
-        if (const auto *apply = dyn_cast<PredicateApply>(expr.get())) {
-          if (apply->getPredicate()->getKind() == Constraint::Kind::SameTypes) {
-            auto arguments = apply->arguments();
-            if (llvm::is_contained(arguments, namedValue.name)) {
-              for (const auto &overloadKey : op->overloadKeys) {
-                StringRef name =
-                    overloadKey.kind == OverloadKey::Result
-                        ? op->results[overloadKey.index].name
-                        : op->arguments[overloadKey.index].name;
-                if (llvm::is_contained(arguments, name))
-                  return false;
-              }
-            }
-          }
-        }
-      }
-
-      return true;
-    };
-
-    for (const auto &result : llvm::enumerate(op->results)) {
-      if (needsOverloadKey(result.value())) {
-        OverloadKey key;
-        key.kind = OverloadKey::Result;
-        key.index = result.index();
-        op->overloadKeys.push_back(key);
-      }
-    }
-    for (const auto &arg : llvm::enumerate(op->getFullArguments())) {
-      if (needsOverloadKey(arg.value())) {
-        OverloadKey key;
-        key.kind = OverloadKey::Argument;
-        key.index = arg.index();
-        op->overloadKeys.push_back(key);
-      }
-    }
-
-    dialectIt->second->operations.push_back(std::move(op));
-  }
-}
-
-std::unique_ptr<PredicateExpr> DialectsContext::parsePredicateExpr(DagInit *dag) {
-  Record *op = dag->getOperatorAsDef({});
-
-  if (op->getName() == "and" || op->getName() == "or" ||
-      op->getName() == "not") {
-    PredicateExpr::Kind kind;
-    if (op->getName() == "and")
-      kind = PredicateExpr::Kind::And;
-    else if (op->getName() == "or")
-      kind = PredicateExpr::Kind::Or;
-    else
-      kind = PredicateExpr::Kind::Not;
-
-    SmallVector<std::unique_ptr<PredicateExpr>> arguments;
-    for (auto [arg, argName] : llvm::zip(dag->getArgs(), dag->getArgNames())) {
-      if (argName) {
-        report_fatal_error(
-            Twine("Logical expression has named arguments: ") 
-                + dag->getAsString());
-      }
-
-      if (!isa<DagInit>(arg)) {
-        report_fatal_error(
-            Twine("Logical expression is missing argument or has non-dag argument: ")
-                + dag->getAsString());
-      }
-
-      arguments.push_back(parsePredicateExpr(cast<DagInit>(arg)));
-    }
-
-    if (kind == PredicateExpr::Kind::Not && arguments.size() != 1) {
-      report_fatal_error(Twine("Logical 'not' must have exactly 1 argument: ")
-                             + dag->getAsString());
-    }
-
-    return std::make_unique<PredicateLogic>(kind, arguments);
-  }
-
-  // It's an application of a predicate.
-  Constraint *predicate = getConstraint(op);
-  SmallVector<std::string> arguments;
-
-  for (auto [arg, argName] : llvm::zip(dag->getArgs(), dag->getArgNames())) {
-    if (!isa<UnsetInit>(arg)) {
-      report_fatal_error(
-          Twine("Predicate application has an argument object: ")
-              + dag->getAsString());
-    }
-
-    if (!argName) {
-      report_fatal_error(
-          Twine("Predicate application is missing argument name: ")
-              + dag->getAsString());
-    }
-
-    arguments.push_back(argName->getValue().str());
-  }
-
-  auto [minArgs, maxArgs] = predicate->getMinMaxArgs();
-  if (arguments.size() < minArgs || arguments.size() > maxArgs) {
-    report_fatal_error(
-        Twine("Predicate application has wrong number of arguments: ")
-            + dag->getAsString());
-  }
-
-  return std::make_unique<PredicateApply>(predicate, arguments);
-}
-
-static std::pair<DialectsContext, Dialect*>
-getSelectedDialect(RecordKeeper& records) {
+static std::pair<GenDialectsContext, GenDialect *>
+getSelectedDialect(RecordKeeper &records) {
   if (g_dialect.empty())
     report_fatal_error(Twine("Must select a dialect using the --dialect option"));
 
-  DialectsContext context;
+  GenDialectsContext context;
   DenseSet<StringRef> dialects;
   dialects.insert(g_dialect);
 
@@ -942,7 +109,7 @@ getSelectedDialect(RecordKeeper& records) {
 
   for (Record* dialectRec : records.getAllDerivedDefinitions("Dialect")) {
     if (dialectRec->getValueAsString("name") == g_dialect) {
-      Dialect* selectedDialect = context.getDialect(dialectRec);
+      GenDialect *selectedDialect = context.getDialect(dialectRec);
       return {std::move(context), selectedDialect};
     }
   }
@@ -1121,7 +288,7 @@ class Builder;
 }
 
 void llvm_dialects::genDialectDefs(raw_ostream& out, RecordKeeper& records) {
-  auto [dialectsContext, dialect] = getSelectedDialect(records);
+  auto [GenDialectsContext, dialect] = getSelectedDialect(records);
 
   emitHeader(out);
 
@@ -1335,8 +502,9 @@ void llvm_dialects::genDialectDefs(raw_ostream& out, RecordKeeper& records) {
       resultTypeName = resultNames[0];
     } else {
       assert(op.results.size() <= 1);
-      Constraint *theResultType =
-          op.results.empty() ? dialectsContext.getVoidTy() : op.results[0].type;
+      Constraint *theResultType = op.results.empty()
+                                      ? GenDialectsContext.getVoidTy()
+                                      : op.results[0].type;
       resultTypeName = typeBuilder.build(theResultType);
     }
